% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cev.R
\name{cevParameterCheck}
\alias{cevParameterCheck}
\title{Input error handling for cev model}
\usage{
cevParameterCheck(parameters)
}
\arguments{
\item{parameters}{vector of three numbers, drift, volatility and cev}
}
\value{
error or null
}
\description{
Input error handling for cev model
}
