% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominantAssetTheorem.R
\name{dat_strategy}
\alias{dat_strategy}
\title{Find the dominant asset via the Kelly-criterion}
\usage{
dat_strategy(r)
}
\arguments{
\item{r}{the periodic arithmetic returns of the assets}
}
\value{
matrix
}
\description{
{Returns in the column index of the dominant asset in a
given dataset of periodic arithemetic returns.}
}
\details{
{This will reutrn an error if no dominant asset is found.}
}
