% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominantAssetTheorem.R
\name{dominant_asset_matrix}
\alias{dominant_asset_matrix}
\title{Estimate the expectation E((1+R_j)/(1+Ri)) for every pair of assets}
\usage{
dominant_asset_matrix(r)
}
\arguments{
\item{r}{the periodic arithmetic returns of the assets}
}
\value{
matrix
}
\description{
{This is simply a sample average of the ratio of returns
between each asset in the data-set.}
}
\details{
{If the row entries are all less than one (except for the diagonal), then
this is the dominant asset.}
}
