% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryWagers.R
\name{entropyBinary}
\alias{entropyBinary}
\title{Growth rate for Kelly fraction for arbitrary binary (non-random) payoffs}
\usage{
entropyBinary(p, a, b)
}
\arguments{
\item{p}{true chance of the event being bet on}

\item{a}{dollars to win}

\item{b}{dollars to risk}
}
\value{
numeric
}
\description{
{The growth rate or entropy of the Kelly strategy}
}
