% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cev.R
\name{entropyCEV}
\alias{entropyCEV}
\title{Optimal growth-rate under CEV diffusions}
\usage{
entropyCEV(t, spot, rate, parameters)
}
\arguments{
\item{t}{time-horizon}

\item{spot}{current spot price}

\item{rate}{risk-free rate}

\item{parameters}{vector of drift, volatility and cev term.}
}
\value{
numeric
}
\description{
{Log return under Kelly-fraction in the CEV model.}
}
