% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteTimeModels.R
\name{entropyDTFM}
\alias{entropyDTFM}
\title{Log-growth n for discrete-time financial market models}
\usage{
entropyDTFM(distr, param, rate = 0, h = 10^-4)
}
\arguments{
\item{distr}{distribution name for the (daily) arithmetic returns}

\item{param}{the parameters of the distribution}

\item{rate}{the risk-neutral rate of the money-market account}

\item{h}{step size for numeric blow-ups}
}
\value{
numeric
}
\description{
{Approximates the integral in the log utility problem. This
is a new interface for \code{entropy_integral}, which will be deprecated and removed.}
}
