% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricBrownianMotion.R
\name{entropyGBM}
\alias{entropyGBM}
\title{The optimal log-growth rate under univariate GBM}
\usage{
entropyGBM(drift, volat, rate = 0)
}
\arguments{
\item{drift}{the mean drift rate of the GBM}

\item{volat}{the volatility of the GBM}

\item{rate}{the risk-free rate of the bond}
}
\value{
numeric
}
\description{
{The optimal log growth rate under the Kelly-criterion for
geometric Brownian motion.}
}
