% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heston.R
\name{entropyHeston}
\alias{entropyHeston}
\title{Optimized log-growth for Heston dynamics}
\usage{
entropyHeston(v0, tt, param, rate = 0, resolution = c(100, 100))
}
\arguments{
\item{v0}{current volatility state}

\item{tt}{time-horizon}

\item{param}{the vector \code{c(kappa, theta, xi, mu)}}

\item{rate}{the risk-free rate}

\item{resolution}{the time and space interval resolution}
}
\value{
numeric
}
\description{
{Compute the log-growth under the Kelly-criterion
for the stochastic Heston volatility model.}
}
