% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itoProcess.R
\name{entropyItoProcess}
\alias{entropyItoProcess}
\title{Log-growth rate under optimal fraction invested in Ito process}
\usage{
entropyItoProcess(t, s, dynamics, rate = 0, resolution = c(200, 200))
}
\arguments{
\item{t}{length of time}

\item{s}{current spot price}

\item{dynamics}{list of drift and volatility function}

\item{rate}{risk-free rate}

\item{resolution}{the time and space resolution for the PDE solver}
}
\value{
numeric
}
\description{
{Compute the entropy rate as a function of the fraction. This
involves solving a Feynman-Kac PDE with running cost the sum of the risk-free rate
and half the square of the market price of risk.}
}
