% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpDiffusion.R
\name{entropyJumpDiffusion}
\alias{entropyJumpDiffusion}
\title{Entropy for jump diffusion log optimal control}
\usage{
entropyJumpDiffusion(a, t, x, mu, rate, volat, lambda, distr, jump_param)
}
\arguments{
\item{a}{control fraction}

\item{t}{time at start}

\item{x}{bankroll}

\item{mu}{drift rate}

\item{rate}{bond rate}

\item{volat}{volatility}

\item{lambda}{mean number of jumps per year}

\item{distr}{distribution of jump size}

\item{jump_param}{parameters of above distribution}
}
\value{
numeric
}
\description{
Entropy for jump diffusion log optimal control
}
