% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDiffusion.R
\name{entropyMixtureDiffusion}
\alias{entropyMixtureDiffusion}
\title{Compute the log-growth rate under optimally controlled mixture diffusions}
\usage{
entropyMixtureDiffusion(t, s, spot, rate, parameters)
}
\arguments{
\item{t}{time horizon}

\item{s}{current price}

\item{spot}{initial price}

\item{rate}{risk-free rate}

\item{parameters}{parameters of mixture}
}
\value{
numeric
}
\description{
{A special case wrapper to \code{entropyItoProcess}. Compute the entropy rate as a function of the fraction. This
involves solving a Feynman-Kac PDE with running cost the sum of the risk-free rate
and half the square of the market price of risk.}
}
