% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandemMoneyline.R
\name{entropyTandemMoneyline}
\alias{entropyTandemMoneyline}
\title{Entropy-growth for log-optimal tandem moneyline bets}
\usage{
entropyTandemMoneyline(p, a, b, u, v)
}
\arguments{
\item{p}{true probability of favorite winning}

\item{a}{winnings if favorite wins}

\item{b}{wager to throw down for favorite \eqn{-b}}

\item{u}{winnings if underdog wins \eqn{+u}}

\item{v}{wager to throw down for underdog}
}
\value{
numeric
}
\description{
{The entropy-growth for the log-optimal bet on two wagers for favorite
vs underdog team to win.}
}
