% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpDiffusion.R
\name{jump_entropy_term}
\alias{jump_entropy_term}
\title{Jump entropy term}
\usage{
jump_entropy_term(a, m, r, v, lambda, distr, jump_param)
}
\arguments{
\item{a}{fraction input}

\item{m}{drift rate}

\item{r}{bond rate}

\item{v}{volatility}

\item{lambda}{mean number of jumps per year}

\item{distr}{distribution of jump sizes}

\item{jump_param}{parameters for distribution}
}
\value{
numeric
}
\description{
Jump entropy term
}
