% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpDiffusion.R
\name{jump_value_integral}
\alias{jump_value_integral}
\title{Jump integral term for log utility}
\usage{
jump_value_integral(a, distr = "norm", jump_param, nstd = 4)
}
\arguments{
\item{a}{control variate}

\item{distr}{distribution of the jump-sizes}

\item{jump_param}{parameters for distribution a list}

\item{nstd}{number of standard deviations for integral region}
}
\value{
numeric
}
\description{
Jump integral term for log utility
}
