% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryWagers.R
\name{kellyBinary}
\alias{kellyBinary}
\title{The optimal bet for a binary wager}
\usage{
kellyBinary(p, a, b)
}
\arguments{
\item{p}{true chance of the event being bet on}

\item{a}{dollars to win}

\item{b}{dollars to risk}
}
\value{
numeric
}
\description{
{The classical Kelly fraction that optimizes long-term log-growth.}
}
