% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cev.R
\name{kellyCEV}
\alias{kellyCEV}
\title{Optimal fraction under CEV diffusions}
\usage{
kellyCEV(spot, rate, parameters)
}
\arguments{
\item{spot}{current spot price}

\item{rate}{risk-free rate}

\item{parameters}{vector of drift, volatility and cev term.}
}
\value{
numeric
}
\description{
{Generalization of Kelly-fraction to CEV model.}
}
