% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteTimeModels.R
\name{kellyDTFM}
\alias{kellyDTFM}
\title{Kelly-criterion for discrete-time financial market models}
\usage{
kellyDTFM(distr, param, rate = 0, h = 10^-4)
}
\arguments{
\item{distr}{distribution name for the (daily) arithmetic returns}

\item{param}{the parameters of the distribution}

\item{rate}{the risk-neutral rate of the money-market account}

\item{h}{step size to avoid numeric blow-ups}
}
\value{
numeric
}
\description{
{Approximates the integral in the log utility problem and
then calls uniroot or Newton-Raphson solver to maximize it.}
}
