% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuropeanOptionKelly.R
\name{kellyEuroBS1}
\alias{kellyEuroBS1}
\title{Kelly-fraction for European options under Black-Scholes dynamics}
\usage{
kellyEuroBS1(mu, volat, rate, spot, premium, delta)
}
\arguments{
\item{mu}{real drift of the underlying stock price}

\item{volat}{volatility of the stock price and volatility used for the option pricing}

\item{rate}{the risk-neutral rate}

\item{spot}{the underlying spot price}

\item{premium}{the premium paid for the option}

\item{delta}{the delta of the option, rate of change with respect to spot-price}
}
\value{
numeric
}
\description{
{A Kelly-fraction generalization for trading a portfolio consiting of
a European option priced under Black-Scholes dynamics with an assumed real-volatility,
and a cash-account or bond at the risk-neutral rate. TWo functions are provided in this package,
one for computing it specified by the exact values, and one computed via specifying the
option via strike, expiry, etc and computing the option value and delta via a pricing method.}
}
