% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuropeanOptionKelly.R
\name{kellyEuroBS2}
\alias{kellyEuroBS2}
\title{Kelly-fraction for European options under Black-Scholes dynamics}
\usage{
kellyEuroBS2(strike, expiry, mu, volat, rate, spot, type, N = 100, M = 100)
}
\arguments{
\item{strike}{strike price of the option}

\item{expiry}{time to maturity of option in trading years}

\item{mu}{real drift of the underlying stock price}

\item{volat}{volatility of the stock price and volatility used for the option pricing}

\item{rate}{the risk-neutral rate}

\item{spot}{the underlying spot price}

\item{type}{type of option, "call" or "put"}

\item{N}{number of sub-intervals in time-grid}

\item{M}{number of sub-intervals in space-grid}
}
\value{
numeric
}
\description{
{A Kelly-fraction generalization for trading a portfolio consiting of
a European option priced under Black-Scholes dynamics with an assumed real-volatility,
and a cash-account or bond at the risk-neutral rate. TWo functions are provided in this package,
one for computing it specified by the exact values, and one computed via specifying the
option via strike, expiry, etc and computing the option value and delta via a pricing method.}
}
