% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itoProcess.R
\name{kellyItoProcess}
\alias{kellyItoProcess}
\title{Kelly-fraction for Ito process with coefficient functions of time and space.}
\usage{
kellyItoProcess(t, s, dynamics, rate = 0, FUN = FALSE)
}
\arguments{
\item{t}{current time, numeric}

\item{s}{current price, numeric greater than zero}

\item{dynamics}{list of drift and volatility coefficient functions of \code{(t, x)}}

\item{rate}{risk-free rate of return}

\item{FUN}{boolean default to false to return  numeric, true to return function}
}
\value{
numeric or function
}
\description{
{The generalized Kelly-fraction for space-time dependent coefficient functions
in Ito processes.}
}
