% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpDiffusion.R
\name{kellyJumpDiffusion}
\alias{kellyJumpDiffusion}
\title{log optimal criterion under jump-diffusions}
\usage{
kellyJumpDiffusion(mu, rate, volat, lambda, distr, jump_param)
}
\arguments{
\item{mu}{drift}

\item{rate}{discounting rate}

\item{volat}{volatility}

\item{lambda}{mean rate of jumps}

\item{distr}{distribution of jump sizes}

\item{jump_param}{parameters of distribution}
}
\value{
numeric
}
\description{
log optimal criterion under jump-diffusions
}
