% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpDiffusion.R
\name{kellyJumpDiffusionFP}
\alias{kellyJumpDiffusionFP}
\title{log optimal criterion under jump-diffusions via fixed-point iterations}
\usage{
kellyJumpDiffusionFP(
  mu,
  rate,
  volat,
  lambda,
  distr,
  jump_param,
  iterations = 100
)
}
\arguments{
\item{mu}{drift}

\item{rate}{discounting rate}

\item{volat}{volatility}

\item{lambda}{mean rate of jumps}

\item{distr}{distribution of jump sizes}

\item{jump_param}{parameters of distribution}

\item{iterations}{number of iterations in fixed-point scheme}
}
\value{
numeric
}
\description{
log optimal criterion under jump-diffusions via fixed-point iterations
}
