% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton.R
\name{kellyMerton}
\alias{kellyMerton}
\title{Kelly-criterion under Merton's jump diffusion}
\usage{
kellyMerton(param, rate = 0, iterations = 500, subdivisions = 500, tol = 10^-6)
}
\arguments{
\item{param}{vector of parameters defining the jump-diffusion dynamics. See details}

\item{rate}{the risk-free rate, or money-market account interest rate}

\item{iterations}{number of iterations to use in the Newton-Raphson method for finding the optimal control as a root}

\item{subdivisions}{number of subdivisions to use in numerical integrations}

\item{tol}{tolerance for the zero to be found.}
}
\value{
list
}
\description{
{Compute the optimal allocation fraction under Merton's jump diffusion dynamics by
root finding via Newton-Raphson's method. }
}
