% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDiffusion.R
\name{kellyMixtureDiffusion}
\alias{kellyMixtureDiffusion}
\title{Kelly-criterion for mixture diffusions}
\usage{
kellyMixtureDiffusion(t, s, spot, rate, param, FUN = FALSE)
}
\arguments{
\item{t}{current time}

\item{s}{current price}

\item{spot}{initial price}

\item{rate}{the discounting rate/money-market account return}

\item{param}{matrix with rows probs, mus, sigmas for the components}

\item{FUN}{boolean whether to return numeric or function}
}
\value{
numeric
}
\description{
{Straightforward generalization of the classic Kelly-fraction for
functions of time and space, specialized to mixture coefficients. A wrapper to \code{kellyItoProcess}.}
}
