% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricBrownianMotion.R
\name{kellyPortfolioGBM}
\alias{kellyPortfolioGBM}
\title{Kelly portfolio in continuous time for GBM market}
\usage{
kellyPortfolioGBM(drift, Sigma, rate = 0, restraint = 1, direction = "long")
}
\arguments{
\item{drift}{the drift vector of the GBMs}

\item{Sigma}{the covariance matrix of the GBMs}

\item{rate}{the rate earned on cash in a money market account, etc}

\item{restraint}{percentage of wealth to bankroll}

\item{direction}{the direction of the bet: long or short}
}
\value{
list of growth rate and optimal portfolio
}
\description{
{A wrapper to \code{solve.QP} for computing optimal portfolios
under a market of correlated geometric Brownian motions.}
}
\details{
{The drift of the log-returns and covariance must be passed.}
}
