% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpDiffusion.R
\name{kjdf_root}
\alias{kjdf_root}
\title{Function to find root of for log optimal criterion under jump-diffusions}
\usage{
kjdf_root(a, mu, rate, volat, lambda, distr, jump_param)
}
\arguments{
\item{a}{control variate}

\item{mu}{drift}

\item{rate}{discounting rate}

\item{volat}{volatility}

\item{lambda}{mean rate of jumps}

\item{distr}{distribution of jump sizes}

\item{jump_param}{parameters of distribution}
}
\value{
numeric
}
\description{
Function to find root of for log optimal criterion under jump-diffusions
}
