% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpDiffusion.R
\name{mgfdkou}
\alias{mgfdkou}
\title{Moment generating function for Displaced Kou distribution}
\usage{
mgfdkou(t, prob, alpha, beta, ku, kd)
}
\arguments{
\item{t}{mgf argument}

\item{prob}{probability of mixtures}

\item{alpha}{mean jump size up}

\item{beta}{mean jump size down}

\item{ku}{displacement up}

\item{kd}{displacement down}
}
\value{
numeric
}
\description{
Moment generating function for Displaced Kou distribution
}
