% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpDiffusion.R
\name{mgfkou}
\alias{mgfkou}
\title{Moment generating function for Displaced Kou distribution}
\usage{
mgfkou(t, prob, alpha, beta)
}
\arguments{
\item{t}{mgf argument}

\item{prob}{probability of mixtures}

\item{alpha}{mean jump size up}

\item{beta}{mean jump size down}
}
\value{
numeric
}
\description{
Moment generating function for Displaced Kou distribution
}
