% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpDiffusion.R
\name{mgfnorm}
\alias{mgfnorm}
\title{Moment-generating function of a Gaussian RV}
\usage{
mgfnorm(t, mean, sd)
}
\arguments{
\item{t}{mgf argument}

\item{mean}{mean of the distribution}

\item{sd}{standard deviation}
}
\value{
numeric
}
\description{
Moment-generating function of a Gaussian RV
}
