% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpDiffusion.R
\name{mgfunif}
\alias{mgfunif}
\title{Moment-generating function of a uniform RV}
\usage{
mgfunif(t, min, max)
}
\arguments{
\item{t}{mgf argument}

\item{min}{minimum of interval}

\item{max}{maximum of interval}
}
\value{
numeric
}
\description{
Moment-generating function of a uniform RV
}
