% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDiffusion.R
\name{mixtureParameterCheck}
\alias{mixtureParameterCheck}
\title{Input error handling for mixture model}
\usage{
mixtureParameterCheck(parameters)
}
\arguments{
\item{parameters}{three rows containing probabilities, drifts, volatilities}
}
\value{
error or null
}
\description{
Input error handling for mixture model
}
