% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalFractions.R
\name{optimalBinary}
\alias{optimalBinary}
\title{The optimal bet for a binary wager}
\usage{
optimalBinary(chance, win, risk)
}
\arguments{
\item{chance}{true chance of event}

\item{win}{dollars to win}

\item{risk}{dollars to risk}
}
\value{
numeric
}
\description{
{The classical Kelly fraction that optimizes long-term log-growth.}
}
