% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cev.R
\name{optimalCEV}
\alias{optimalCEV}
\title{Simulate log-optimal strategy on CEV diffusions}
\usage{
optimalCEV(bankroll, t, spot, rate, parameters)
}
\arguments{
\item{bankroll}{initial bankroll to invest}

\item{t}{time horizon to trade over}

\item{spot}{the initial stock price}

\item{rate}{the return of the bond}

\item{parameters}{vector of cev parameters}
}
\value{
data.frame of solution
}
\description{
{Simulate log-optimal strategy for cev diffusion prices. A wrapper
to \code{optimalItoProcess}.}
}
\details{
{parameters must be a vector of drift, volatility and cev.}
}
