% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricBrownianMotion.R
\name{optimalGBM}
\alias{optimalGBM}
\title{Simulate log-optimal strategy on mixture diffusions}
\usage{
optimalGBM(bankroll, t, spot, rate, parameters)
}
\arguments{
\item{bankroll}{initial bankroll to invest}

\item{t}{time horizon to trade over}

\item{spot}{the initial stock price}

\item{rate}{the return of the bond}

\item{parameters}{drift and volatility}
}
\value{
data.frame of solution
}
\description{
{Simulate log-optimal strategy for prices under geometric Brownian motion. A wrapper
to \code{optimalItoProcess}.}
}
\details{
{The parameters must be a vector of drift and volatility.}
}
