% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heston.R
\name{optimalHeston}
\alias{optimalHeston}
\title{Simulate a sample-path of the Kelly-portfolio under Heston dynamics}
\usage{
optimalHeston(x0, s0, v0, tt, param, rate = 0, n = 1000, plotG = TRUE)
}
\arguments{
\item{x0}{initial wealth to invest}

\item{s0}{initial spot price of the stock}

\item{v0}{initial (hidden) stochastic volatility level}

\item{tt}{time horizon to simulate a path over}

\item{param}{vector of \code{kappa, theta, xi, mu, rho}}

\item{rate}{risk-free rate of cash-account, money-market account, or bond}

\item{n}{number of time sub-intervals in sample-path}

\item{plotG}{boolean for plotting the portfolio in the function body}
}
\value{
data.frame of time, volatility, stock-price, and portfolio-value
}
\description{
{A basic Euler-Maruyama scheme for simulating a sample path of
a stochastic volatility price model and its corresponding Kelly-strategy.}
}
