% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itoProcess.R
\name{optimalItoProcess}
\alias{optimalItoProcess}
\title{Simulate log-optimal strategy for a given Ito process}
\usage{
optimalItoProcess(bankroll, t, spot, rate, dynamics, n = 1000)
}
\arguments{
\item{bankroll}{initial bankroll to invest}

\item{t}{time horizon to trade over}

\item{spot}{the initial stock price}

\item{rate}{the return of the bond}

\item{dynamics}{list of drift and volatility function of \code{(t, s)}}

\item{n}{number of time-steps to use in sample-path simulation}
}
\value{
data.frame of solution
}
\description{
{Simulate log-optimal strategy for prices driven by an Ito process}
}
\details{
{The argument \code{dynamics} must be a named list of functions \code{drift(t,x)} and
\code{diffusion(t, x)}.}
}
