% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton.R
\name{optimalMertonPath}
\alias{optimalMertonPath}
\title{Euler-Maruyama scheme for Kelly portfolio under Merton's jump diffusion}
\usage{
optimalMertonPath(t, param, rate = 0, n = 1000)
}
\arguments{
\item{t}{maturity to simulate under}

\item{param}{vector of parameters defining the jump-diffusion dynamics. See details}

\item{rate}{the risk-free rate, or money-market account interest rate}

\item{n}{number of time-subintervals to use in the Euler-Maruyama scheme}
}
\value{
list
}
\description{
{Simulate a sample path of \eqn{\log X_t} of the Kelly-portfolio
under Merton's jump diffusion dynamics, via the Euler-Maruyama scheme}
}
