% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDiffusion.R
\name{optimalMixtureDiffusion}
\alias{optimalMixtureDiffusion}
\title{Simulate log-optimal strategy on mixture diffusions}
\usage{
optimalMixtureDiffusion(bankroll, t, spot, rate, param)
}
\arguments{
\item{bankroll}{initial bankroll to invest}

\item{t}{time horizon to trade over}

\item{spot}{the initial stock price}

\item{rate}{the return of the bond}

\item{param}{matrix of parameters of mixture, see details}
}
\value{
data.frame of solution
}
\description{
{Simulate log-optimal strategy for mixture diffusion prices. A wrapper
to \code{optimalItoProcess}.}
}
\details{
{The matrix must contain a row of probabilities, a row of means, and
a row of standard deviations.}
}
