% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricBrownianMotion.R
\name{portfolioGBM}
\alias{portfolioGBM}
\title{Kelly portfolio in continuous time for GBM market}
\usage{
portfolioGBM(
  log_ret,
  rate = 0,
  restraint = 1,
  direction = "long",
  sample_mean = NULL,
  sample_cov = NULL
)
}
\arguments{
\item{log_ret}{the data-set of (daily) log-returns. Assumes it is a matrix with colnames of tickers}

\item{rate}{the rate earned on cash in a money market account, etc}

\item{restraint}{percentage of wealth to bankroll}

\item{direction}{the direction of the bet: long or short}

\item{sample_mean}{the mean vector, optional}

\item{sample_cov}{the covariance matrix, optional}
}
\value{
list of growth rate and optimal portfolio
}
\description{
{For continuous time market of GBMs, the log optimal portfolio is given by a quadratic
programming problem.}
}
\details{
{The function estimates the entire sample set for mean returns and covariance and then runs the optimization routine by calling solve.qp from quadprog package.}
}
