% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteTimeModels.R
\name{simulateDTFM}
\alias{simulateDTFM}
\title{Simulate log-optimal strategy for DTFM}
\usage{
simulateDTFM(n, spot, bankroll, distr, param, rate = 0, plotG = TRUE)
}
\arguments{
\item{n}{number of days to simulate}

\item{spot}{initial stock price}

\item{bankroll}{initial bankroll}

\item{distr}{distribution of daily arithmetic returns}

\item{param}{parameters of the distribution}

\item{rate}{risk-neutral rate}

\item{plotG}{boolean for plotting on call}
}
\value{
list
}
\description{
{Simulates daily stock prices under the model and implements the strategy.
Assuming no trading costs, etc. Returns mean and total growth plus pnl.}
}
