% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachingWealthGoal.R
\name{wealthGoalControl}
\alias{wealthGoalControl}
\title{The optimal control for reaching a wealth goal by a terminal date}
\usage{
wealthGoalControl(t, x, b, maturity, mu, volat, rate)
}
\arguments{
\item{t}{current time}

\item{x}{initial wealth level}

\item{b}{target wealth level}

\item{maturity}{terminal date to hit the goal by}

\item{mu}{the drift of the stock}

\item{volat}{the volatility of the stock}

\item{rate}{the risk-free rate from a cash-account, bond, etc}
}
\value{
numeric, matrix, vector
}
\description{
{The optimal control for maximizing the probability of the terminal wealth
of a GBM-cash portfolio being greater than a fixed level, conditional on the initial wealth.}
}
