% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryWagers.R
\name{optimalBinary}
\alias{optimalBinary}
\title{Simulate Kelly strategy for generalized coin toss}
\usage{
optimalBinary(bankroll, p, a, b, trials = 100)
}
\arguments{
\item{bankroll}{initial bankroll to bet with}

\item{p}{probability of winning each trial}

\item{a}{payout on winning toss on top of wager}

\item{b}{wager lost on each trial on losing toss}

\item{trials}{number of trials to simulate}
}
\value{
vector
}
\description{
{Simulating the generalized coin tossing gambling game.}
}
\details{
{Start off with a given value of wealth and bet the Kelly fraction each round.}
}
