% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itoProcess.R
\name{optimalItoProcess}
\alias{optimalItoProcess}
\title{Simulate log-optimal strategy for a given Ito process}
\usage{
optimalItoProcess(bankroll, t, spot, rate, dynamics)
}
\arguments{
\item{bankroll}{initial bankroll to invest}

\item{t}{time horizon to trade over}

\item{spot}{the initial stock price}

\item{rate}{the return of the bond}

\item{dynamics}{list of drift and volatility function of \code{(t, s)}}
}
\value{
data.frame of solution
}
\description{
{Simulate log-optimal strategy for prices driven by an Ito process}
}
\details{
{...}
}
